pyproject.toml
[tool.poetry]
name = "databooks"
version = "0.0.0-dev"  # dynamically set in CI
description = "A CLI tool to resolve git conflicts and remove metadata in notebooks."
authors = ["Murilo Cunha <murilo@dataroots.io>"]
license = "MIT"
readme = "README.md"
homepage = "https://datarootsio.github.io/databooks/"
repository = "https://github.com/datarootsio/databooks/"
keywords = ["jupyter-notebooks", "cli"]

[tool.poetry.scripts]
databooks = "databooks.cli:app"

[tool.poetry.dependencies]
python = "^3.7"
typer = ">=0.4.0,<1.0.0"
rich = "^12.6.0"
pydantic = "^2.3"
GitPython = "^3.1.24"
tomli = "^2.0.1"
typing-extensions = "^4.0.1"


[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"
pytest = "^7.2.0"
pytest-cov = "^3.0.0"
mike = "^1.1.2"
mkdocstrings = "^0.17.0"
mkdocs-material = "^8.0.5"
mkdocs-autorefs = "^0.3.1"
mkdocs-coverage = "^0.2.4"
Jinja2 = "3.0.3"  # older versions have breaking changes for `mike`

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Devtools not in environment, but used in `pre-commit`
[tool.black]
include = "(databooks/|tests/)"
extend-exclude = "tests/files/"

[tool.mypy]
files = ["databooks", "tests"]
show_error_codes = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true
warn_no_return = false

[tool.isort]
skip_glob = ["docs/*"]
profile = "black"
